p1/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
sqlalchemy==2.0.22
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
python-multipart==0.0.6
google-generativeai>=0.8.0
httpx==0.25.0